cleaner-bot/pyproject.toml
[project]
name = "cleanerbot"
version = "0.1.0"
description = "Telegram bot that cleans noise from .docx and .xlsx quiz files."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0",
  "python-dotenv>=1.0.0",
  "openpyxl>=3.1.0",
  "python-docx>=1.1.0",
]

[project.scripts]
cleanerbot = "cleanerbot.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
